src/latexify/requirements.txt
pypdf>=3.9.0
pdf2image>=1.17.0
pillow>=10.0.0
pytesseract>=0.3.10
pytest>=8.2.0
requests>=2.32.0
torch>=2.4.0  # Upgrade for vLLM
torchvision>=0.19.0
transformers>=4.46.0
pydantic-settings>=2.0.0
datasets>=2.19.0
accelerate>=0.33.0
sentencepiece>=0.1.99
pydantic==2.9.2 # Pin compatible
huggingface-hub>=0.25.0
einops>=0.7.0
timm>=1.0.0
duckduckgo-search>=5.3.0
PyMuPDF>=1.24.0
sympy>=1.12
latex2sympy2>=1.0.1
jinja2>=3.1.0
llama-cpp-python==0.2.79
texoutparse>=0.4.0
protobuf>=4.24.0
vllm==0.6.0 # Pin vLLM to stable version
surya-ocr>=0.3.0
matplotlib>=3.8.0
reportlab>=4.0.0
diffusers>=0.29.0
paddlepaddle>=2.6.0
paddleocr>=2.7.0
ultralytics>=8.2.0
opencv-python-headless>=4.9.0
docling[vlm,tesserocr]>=1.0.0
nougat-ocr>=0.1.17
pix2tex>=0.1.0